Rough sketch for bootstrap modal architecture
