+++ src/components/index.ts
@@ -12,9 +12,9 @@ export * from "./logger/logger";
 export * from "./axes/axes";
 export * from "./verification-grid/verification-grid";
 export * from "./verification-grid-tile/verification-grid-tile";
+export * from "./bootstrap-modal/bootstrap-modal";
 export * from "./info-card/info-card";
 export * from "./data-source/data-source";
-export * from "./verification-grid/help-dialog";
 export * from "./decision/decision";
 export * from "./decision/classification/classification";
 export * from "./decision/verification/verification";
