--- src/components/verification-grid/slides/abstractSlide.ts
@@ -1,3 +1,0 @@
-import { LitElement } from "lit";
-
-export abstract class AbstractSlide extends LitElement {}
